+++ src/lib/types/errors/error-response-body.type.ts
@@ -1,0 +1,6 @@
+export type ErrorResponseBody = {
+  message: string;
+  path: string;
+  statusCode: number;
+  timestamp: string;
+};